+++ registry/lib/components/video/player/custom-auto-play/AutoplayActionType.ts
@@ -1,0 +1,9 @@
+/** 自动连播行为类型 */
+export enum AutoplayActionType {
+  /** 自动判断，参考《传统连播模式》组件 */
+  AUTO = '自动',
+  /** 禁用自动连播 */
+  DISABLE = '禁用',
+  /** 总是自动连播 */
+  ALWAYS = '总是',
+}
