+++ registry/lib/components/video/player/custom-auto-play/handlers/BangumiAutoplayHandler.ts
@@ -1,0 +1,32 @@
+import { matchUrlPattern } from '@/core/utils'
+import { bangumiUrls } from '@/core/utils/urls'
+import { AutoplayActionType } from '../AutoplayActionType'
+import { BaseAutoplayHandler } from './BaseAutoplayHandler'
+
+/** 自动连播处理器-番剧 */
+export class BangumiAutoplayHandler extends BaseAutoplayHandler {
+  type = '番剧'
+
+  async match() {
+    return bangumiUrls.some(url => matchUrlPattern(url))
+  }
+
+  protected override isLastSequentialNumber(): boolean {
+    // PV、小剧场没有分P序号，因此统一用元素位置判断
+    const el = document
+      .querySelector('.plp-r img[class^=PlayingIcon_playIcon]')
+      .closest('div[class^=imageListItem_wrap]')
+    return !el.nextElementSibling
+  }
+
+  async shouldAutoplay() {
+    return BaseAutoplayHandler.shouldAutoplayWithAutoHandler(
+      BaseAutoplayHandler.settings.options.bangumiAutoplayAction as AutoplayActionType,
+      () => !this.isLastSequentialNumber(),
+    )
+  }
+
+  async setupAutoPlay(enable: boolean) {
+    await this.setupAutoPlay_Player(enable)
+  }
+}
